解决iphone真机Crash的bug
原因是内存池造成的。iphone内存分配必须对齐在粒度上，如果没有对齐会报访问错误。

+++ FootBallX/Classes/JsonMemory.h
@@ -46,7 +46,7 @@ protected:
     std::vector<int> m_poolSize;
     
     typedef int BLOCK_HEADER_TYPE;
-    typedef unsigned short UNIT_HEADER_TYPE;
+    typedef int UNIT_HEADER_TYPE;
     
     const int START_BUFFER_SIZE = 32;
